% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_conformal.R
\name{adaptive_conformal}
\alias{adaptive_conformal}
\title{Build prediction sets with respect to a significance level.}
\usage{
adaptive_conformal(
  x_train,
  y_train,
  x_cal,
  y_cal,
  x_test,
  y_test,
  method = "rf",
  epsilon = 0.1,
  feature_selection = FALSE,
  num_features = NULL
)
}
\description{
Build prediction sets with respect to a significance level.
}
