% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_sets.R
\name{predict_sets}
\alias{predict_sets}
\title{Build prediction sets with respect to significance level.}
\usage{
predict_sets(Y_pred, epsilon, unif)
}
\description{
Build prediction sets with respect to significance level.
}
