% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_conformal.R
\name{split_conformal}
\alias{split_conformal}
\title{Basic split conformal method for building prediction sets with respect to significance level.}
\usage{
split_conformal(
  x_train,
  y_train,
  x_cal,
  y_cal,
  x_test,
  y_test,
  method = "rf",
  epsilon = 0.1,
  feature_selection = FALSE,
  num_features = NULL
)
}
\description{
Basic split conformal method for building prediction sets with respect to significance level.
}
