% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_coverage.R
\name{target_coverage}
\alias{target_coverage}
\title{Compute marginal coverage for each target value.}
\usage{
target_coverage(y_test, predicted_set)
}
\description{
Compute marginal coverage for each target value.
}
